fix compile errors in lfs module, guard addon UI on missing module
Ref<T> misuse for non-RefCounted LfsRemoteClient, MIN() cast ambiguity,
HTTPClient API mismatch, and CoreBind namespace fixed across lfs/*.cpp.
plugin.gd now checks ClassDB.class_exists("GitBackend") before wiring up
Git/Dotfiles/LFS tabs so editors without the custom module degrade
gracefully instead of throwing.

+++ lfs/lfs_status_scanner.cpp
@@ -56,7 +56,7 @@ Dictionary LfsStatusScanner::classify_file(const String &project_root, const Str
 		d["status"] = CLEAN;
 		return d;
 	}
-	PackedByteArray prefix = prefix_file->get_buffer(MIN(prefix_file->get_length(), (int64_t)STATUS_PREFIX_SAMPLE_BYTES));
+	PackedByteArray prefix = prefix_file->get_buffer(MIN((int64_t)prefix_file->get_length(), (int64_t)STATUS_PREFIX_SAMPLE_BYTES));
 	prefix_file->close();
 
 	if (LfsPointer::looks_like_pointer(prefix)) {
